move boxdrawer to separate file
